renaming model.py to models.py
