+++ algorithmic_efficiency/workloads/librispeech_deepspeech/librispeech_pytorch/workload.py
@@ -6,13 +6,13 @@ from torch.nn.parallel import DistributedDataParallel as DDP
 from algorithmic_efficiency import param_utils
 from algorithmic_efficiency import spec
 from algorithmic_efficiency.pytorch_utils import pytorch_setup
-from algorithmic_efficiency.workloads.librispeech_conformer.librispeech_pytorch.model import \
+from algorithmic_efficiency.workloads.librispeech_conformer.librispeech_pytorch.models import \
     initialize
 from algorithmic_efficiency.workloads.librispeech_conformer.librispeech_pytorch.workload import \
     LibriSpeechConformerWorkload
-from algorithmic_efficiency.workloads.librispeech_deepspeech.librispeech_pytorch.model import \
+from algorithmic_efficiency.workloads.librispeech_deepspeech.librispeech_pytorch.models import \
     DeepspeechConfig
-from algorithmic_efficiency.workloads.librispeech_deepspeech.librispeech_pytorch.model import \
+from algorithmic_efficiency.workloads.librispeech_deepspeech.librispeech_pytorch.models import \
     DeepspeechEncoderDecoder
 
 USE_PYTORCH_DDP, RANK, DEVICE, N_GPUS = pytorch_setup()
